Create scene with minimum size and set stage min size to the starting size.
As a scene is created with minimum size (set in the FXML file) the stage
starts with the size that is the scene's min size + border size of the 
window. Thus setting the min size of the stage upon it's showing to the
current size is equivalent to not allowing a client to resize below
the min constraints set in the FXML file.